Address review feedback on PR #70
## Correctness

- **FK column type now matches the referenced `id`**. Reviewers flagged
  (5x) that inferred FKs were hardcoded to BIGINT while the target
  entity's `id` column was Integer (e.g. `users.id: Integer`,
  `orders.id: Integer`). Schema now precomputes per-entity
  `{tableName, idFkSqlType}` in `buildEntityRefMap`, respecting
  `db_table` overrides and resolving type aliases. FK columns inherit
  that SQL type. BIGSERIAL (synthetic PK) collapses to BIGINT for FK
  references.
- **Redirect classification narrowed**. 304 Not Modified is not a
  navigational redirect — only 301/302/303/307/308 trigger the
  RedirectResponse branch now.
- **OpenAPI Read schemas also strip sensitive fields**. The Pydantic
  filter was applied but the OpenAPI contract was still exposing
  `password_hash` etc., desyncing the two. Extracted
  `isSensitiveFieldName` to `src/codegen/sensitive-fields.ts` and
  applied in both emit.ts and openapi/components.ts.

## Code quality

- **Dropped `as readonly (string | null)[]` cast** in
  `openapi/schema.ts:makeNullable` — CLAUDE.md bans `as` casts.
  Returns flow via two branches now, TS infers cleanly.
- **Type alias resolution handles OptionType aliases**. `type X =
  Option[Int]` now resolves to `int | None` instead of falling back
  to `str`. No fixture exercises this today but it's correct.
- **B008 suppression in generated `pyproject.toml`**. FastAPI's
  `Depends(...)` in default argument position is evaluated per-request,
  so ruff B008 is a false positive for this codebase. Added
  `extend-ignore = ["B008"]` under `[tool.ruff.lint]`.

## Test hardening

- `migration.test.ts`: count `ck_*` constraint names per table and
  assert the count equals `uniqueChecks.length`, so a regression that
  emits extras alongside the expected names still fails.
- `emit.test.ts` + `services.test.ts`: collapse the two separate
  assertions (`async def shorten(...)` + generic `NotImplementedError`)
  into a single regex that requires the stub to be the body of
  `shorten` specifically, not just present somewhere in the file.

## Verification

- `npx tsc --noEmit` clean
- 1090 / 1090 pass; goldens deterministic (second `-u` no-op)
- All 113 emitted `.py` files `ast.parse` clean
- `npm run fmt:check` clean

+++ test/codegen/golden/expected/auth_service/alembic/versions/001_initial_schema.py
@@ -34,7 +34,7 @@ def upgrade() -> None:
     op.create_table(
         "sessions",
         sa.Column("id", sa.Integer(), primary_key=True, nullable=False),
-        sa.Column("user_id", sa.BigInteger(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
         sa.Column("access_token", sa.Text(), nullable=False),
         sa.Column("refresh_token", sa.Text(), nullable=False),
         sa.Column("access_expires_at", sa.DateTime(timezone=True), nullable=False),
